+++ src/plugins/core/widgets/index.ts
@@ -4,5 +4,4 @@ export { default as Greeting } from './Greeting';
 export { default as GreetingSettings } from './GreetingSettings';
 export { default as Message } from './Message';
 export { default as MessageSettings } from './MessageSettings';
-export { default as Time } from './Time';
-export { default as TimeSettings } from './TimeSettings';
+export * from './time';