+++ src/plugins/core/widgets/time/index.ts
@@ -1,0 +1,2 @@
+export { default as Time } from './Time';
+export { default as TimeSettings } from './TimeSettings';