+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- New analogue mode for the time widget.
+
 ### Fixed
 - Black screens experienced in Firefox private windows on second (and subsequent) new tabs.